+++ app/src/main/java/com/lrdb/la_roja_de_bruselas_kt/LaRojaDeBruselasApp.kt
@@ -1,0 +1,6 @@
+package com.lrdb.la_roja_de_bruselas_kt
+
+import android.app.Application
+
+class LaRojaDeBruselasApp: Application {
+}
